improve some payment steps
