+++ resources/js/components/PaymentStatusComponent.vue
@@ -1,6 +1,6 @@
 <template>
     <div>
-        <div class="p-2 flex gap-8 flex-col justify-center items-center" v-if="isPaymentSuccess && isWaitingDone">
+        <div class="p-2 flex gap-8 flex-col justify-center items-center" v-if="successfulPayment">
             <div class="font-bold text-green-500">
                 <div class="flex flex-col gap-2 w-full items-center">
                     <div class="flex flex-col text-gray-900 gap-2 justify-center items-center p-5">
@@ -11,21 +11,21 @@
                     <div class="mt-5">
                         <div class="flex justify-between">
                             <div>Invoice Number</div>
-                            <div>{{payment.invoice_id}}</div>
+                            <div>{{successfulPayment.invoice_id}}</div>
                         </div>
                         <div class="flex justify-between">
                             <div>Amount</div>
-                            <div>{{payment.amount}}</div>
+                            <div>{{successfulPayment.amount}}</div>
                         </div>
                         <div class="flex justify-between">
                             <div>Paid at</div>
-                            <div>{{payment.created_at}}</div>
+                            <div>{{successfulPayment.created_at}}</div>
                         </div>
                     </div>
                 </div>
             </div>
         </div>
-        <div class="p-2 flex gap-8 flex-col justify-center items-center" v-if="isWaitingDone && isPaymentSuccess!==null && !isPaymentSuccess">
+        <div class="p-2 flex gap-8 flex-col justify-center items-center" v-if="!successfulPayment">
             <div class="flex flex-col gap-2 w-full items-center">
                 <div class="flex flex-col text-gray-900 gap-2 justify-center items-center p-5">
                     <span class="flex justify-center items-center w-10 h-10 bg-red-800 rounded-full text-white"><i class="fa fa-exclamation-circle fa-lg"></i></span>
@@ -39,7 +39,7 @@
                 </div>
             </div>
         </div>
-        <div v-if="isWaitingDone" class="flex justify-end gap-2 p-2">
+        <div class="flex justify-end gap-2 p-2">
             <a href="/home"  class="text-white bg-blue-700 hover:bg-blue-800 focus:ring-4 focus:outline-none focus:ring-blue-300 font-medium rounded-lg text-sm px-5 py-2.5 text-center dark:bg-blue-600 dark:hover:bg-blue-700 dark:focus:ring-blue-800">
                 Continue Shopping
             </a>
@@ -50,57 +50,11 @@
     </div>
 </template>
 <script setup>
-
-import {onMounted, ref} from "vue";
-
 const props = defineProps({
-    paymentId:{
-        type:String,
-        required:true
+    successfulPayment:{
+        type:Object,
+        required:true,
+        default:null
     }
 })
-const emit = defineEmits(['onPaymentResultKnown'])
-
-const isPaymentSuccess = ref(null)
-
-const pollingInterval = 1000; // 1 second
-const totalDuration = 3 * 60 * 1000; // 3 minutes in milliseconds
-let startTime = Date.now();
-const isWaitingDone = ref(false)
-const payment = ref(null)
-
-onMounted(async () => {
-    await pollIfPaymentSuccess()
-})
-
-async  function pollIfPaymentSuccess(){
-    const url = `/api/payments/${props.paymentId}/status`
-    const response = await fetch(url, {
-        method: 'GET',
-        headers: {
-            'Content-Type': 'application/json',
-        }
-    })
-
-    const responseData = await response.json();
-
-    const elapsedTime = Date.now() - startTime;
-    if (elapsedTime < totalDuration && !isWaitingDone.value) {
-        // Continue polling after a certain delay
-        setTimeout(async () => await pollIfPaymentSuccess(), pollingInterval);
-    } else {
-        console.log('Polling completed.');
-    }
-
-    if(responseData.status==="success"){
-        isPaymentSuccess.value = true
-        isWaitingDone.value = true
-        payment.value = responseData.data
-        emit('onPaymentResultKnown',isPaymentSuccess.value)
-    }else if(responseData.status==="failed"){
-        isPaymentSuccess.value = false
-        isWaitingDone.value = true
-        emit('onPaymentResultKnown',isPaymentSuccess.value)
-    }
-}
 </script>
